+++ examples/plot_run_benchmark.py
@@ -21,9 +21,8 @@ try:
     save_file = run_benchmark(
         Benchmark(BENCHMARK_PATH), ['sklearn[liblinear]', 'sklearn[newton-cg]',
                                     'lightning'],
-        dataset_names=['Simulated*n_samples=200,n_features=500*'],
-        objective_filters=[
-            'L2 Logistic Regression[lmbd=1.0]'],
+        dataset_names=['Simulated*n_features=500,n_samples=200*'],
+        objective_filters=['L2 Logistic Regression[lmbd=1.0]'],
         max_runs=100, timeout=20, n_repetitions=15,
         plot_result=False, show_progress=True
     )